+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- \#415 Unable to enter parameters
 
 ## [0.13.15-2]
 - \#413 0.13.15-2022.2 - still not working with Intellij 2022.3